requirements.txt
# Core ML/DL frameworks
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Web Application
streamlit>=1.28.0

# Progress Bar
tqdm>=4.66.0
